Fix executable jar not working (#32)
* Fix executable jar not working

Changed main class FQN in pom.xml, so the jar can be properly built
and launched.

* Add tests for checking proper jar creation

Modified pom to make maven generate sample jar before testing
(during generate-test-sources phase). Having that we can check whether
the file was created at all and if it contains main class FQN
in the manifest.

* Change to use maven-failsafe-plugin instead of creating another jar before package phase

Using failsafe plugin allows to run integration tests. The phase is
executed after packaging so the original jar file can be checked.

* Rename test to fit integration test naming scheme and skip other tests if file is null

The base test checks whether the file read from the file system is not null and
if that is the case the other tests can process. Otherwise they are skipped.