9 rotation of blocks (#56)
* Introduce RotatedBlock

* Unit tests for RotatedBlock

* Unit tests for Playfield

At least one block type must be public to test this feature. My choice is Z-Block as it gives perfect opportunities to investigate rotations.

* Update testng.xml

* Update version to 1.0

* Correct tests - use new Grid class in PlayfieldTest

* Create public MockedBlock in tests

Public MockedBlock can be used in tests outside block package, for example in PlayfieldTest. This remove necessity for public or protected access in the source code.

Co-authored-by: Krzysiek <kjanas@cern.ch>